+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- 2019-12-11: Add MocoOutputGoal, allowing any scalar model output to be used as
+              as a goal.
+
 - 2019-12-11: ActivationCoordinateActuator now provides default activation 
               bounds equal to its control bounds.